Attempt to implement fuji raw conv, add fuji files, add length option for props if type is PTP_TC_UNDEF

+++ src/data.c
@@ -253,7 +253,7 @@ int put_string(unsigned char *data, const char *str) {
 	for (i = 0; i < data[0]; i++)
 		put_16bit_le(data + 1 + 2 * i, str[i]);
 
-	return 1 + (strlen(str) + 1) * 2;
+	return 1 + (int)(strlen(str) + 1) * 2;
 }
 
 char *get_string(unsigned char *data) {
